backend/requirements.txt
fastapi==0.115.0
mangum==0.17.0
pydantic>=2.0.0
psycopg2-binary==2.9.9
pgvector==0.3.2
httpx==0.27.0
boto3==1.35.0
uvicorn==0.30.0